Move files
Move
	TestNanoLoanModel.js
	TestDebtEngine.js
To diaspore folder
